refactor(front): merge shared and components folders
