+++ packages/front/package.json
@@ -23,8 +23,6 @@
         "url": "git+https://github.com/betagouv/datasubvention.git"
     },
     "scripts": {
-        "checkHTML": "site-validator --verbose --url",
-        "checkHTMLLocal": "site-validator --verbose --local --url http://localhost:1234",
         "dev": "nodemon",
         "start": "node ./build/src/index.js",
         "prebuild": "webpack && webpack --config ./webpack.svelte.config.js",
