+++ packages/front/svelte/components/Footer.svelte
