+++ packages/front/svelte/core/SseConnector.js
