+++ packages/front/svelte/resources/subventions/subventions.port.js
@@ -1,7 +1,7 @@
 // Voir si il ne vaux mieux pas passer ça en adapteur
 import { flatenProviderValue } from "../../helpers/providerValueHelper";
-import SSEConnector from "../../shared/SseConnector";
-import Store from "../../shared/Store";
+import SSEConnector from "../../core/SseConnector";
+import Store from "../../core/Store";
 
 class SubventionsPort {
     _getSubventionsConnectedStore(identifier, type) {
