--- packages/front/svelte/shared/Component.core.js
@@ -1,21 +1,0 @@
-export default class ComponentCore {
-    constructor() {
-        this.renderCallback = () => null;
-    }
-
-    onRender(cb) {
-        this.renderCallback = cb;
-    }
-
-    render() {
-        this.renderCallback(this.buildRendererData());
-    }
-
-    destroy() {
-        return;
-    }
-
-    buildRendererData() {
-        return {};
-    }
-}
