+++ packages/front/svelte/views/etablissement/Etablissement.svelte
@@ -2,7 +2,7 @@
     import Button from "../../dsfr/Button.svelte";
     import DataNotFound from "../../components/DataNotFound.svelte";
     import ErrorAlert from "../../components/ErrorAlert.svelte";
-    import InfosLegales from "../../shared/InfosLegales.svelte";
+    import InfosLegales from "../../components/InfosLegales.svelte";
     import TabEtab from "./components/TabEtab.svelte";
     import Spinner from "../../components/Spinner.svelte";
     import etablissementService from "./etablissement.service";
